增加区域Model,调整加载事件位置
