+++ source/model/index.ts
@@ -6,6 +6,7 @@ import { LogisticsModel } from './Logistics';
 import { HotelCanStayingModel } from './HotelCanStaying';
 import { ClinicModel } from './Clinic';
 import { FactoryService } from './services/FactoryService';
+import { AreaModel } from './district/Area';
 
 export const history = new History();
 export const session = new Session();
@@ -14,6 +15,7 @@ export const logistics = new LogisticsModel();
 export const hotelCanStaying = new HotelCanStayingModel();
 export const clinic = new ClinicModel();
 export const FactoryStore = new FactoryService();
+export const area = new AreaModel();
 
 export * from './HTTPService';
 export * from './GitService';
